mvp: wasm target cleanup
