+++ engine/pill_engine/src/resources/texture.rs
@@ -37,6 +37,16 @@ impl Texture {
             renderer_resource_handle: None,
         }
     }
+
+    /// Build a texture from raw image bytes (e.g. `include_bytes!("…png")`).
+    /// Decoded via the `image` crate — format detected from the bytes.
+    pub fn from_bytes(name: &str, texture_type: TextureType, bytes: &[u8]) -> Self {
+        Self::new(
+            name,
+            texture_type,
+            ResourceLoader::Bytes(bytes.to_vec().into_boxed_slice()),
+        )
+    }
 }
 
 impl PillTypeMapKey for Texture {
